Refactoring: migrate to multi sub-modules project (#3)
* refactoring: using submodules

* refactoring: build fonts jar and use it at service
