+++ fonts/build.gradle
@@ -1,0 +1,4 @@
+jar {
+    classifier = null
+    version = null
+}
