+++ fonts/src/main/resources/fonts/fontsfamily1457043146974.xml
@@ -1,0 +1,11 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<fontFamilies>
+    <fontFamily name="Roboto">
+        <normal><![CDATA[fonts/Roboto-Regular.ttf]]></normal>
+        <bold><![CDATA[fonts/Roboto-Bold.ttf]]></bold>
+        <italic><![CDATA[fonts/Roboto-Italic.ttf]]></italic>
+        <boldItalic><![CDATA[fonts/Roboto-BoldItalic.ttf]]></boldItalic>
+        <pdfEncoding><![CDATA[Identity-H]]></pdfEncoding>
+        <pdfEmbedded><![CDATA[true]]></pdfEmbedded>
+    </fontFamily>
+</fontFamilies>
