+++ fonts/src/main/resources/jasperreports_extension.properties
@@ -1,0 +1,2 @@
+net.sf.jasperreports.extension.registry.factory.fonts=net.sf.jasperreports.engine.fonts.SimpleFontExtensionsRegistryFactory
+net.sf.jasperreports.extension.simple.font.families.ireportfamily1457043146974=fonts/fontsfamily1457043146974.xml
