+++ service/src/test/resources/3789_extra_picture.jrxml
