+++ service/src/test/resources/3800_extra_picture.jrxml
