+++ service/src/test/resources/checked.png
