+++ service/src/test/resources/data_1_image.xml
