+++ service/src/test/resources/data_3_images.xml
