+++ service/src/test/resources/data_without_images.xml
