+++ service/src/test/resources/logo.png
