+++ service/src/test/resources/logo1.png
