+++ service/src/test/resources/logo2.png
