+++ service/src/test/resources/report_with_images.jrxml
