+++ service/src/test/resources/report_without_https_images.jrxml
