+++ service/src/test/resources/subreport_single_image.jrxml
