+++ service/src/test/resources/test_with_image.jrxml
