+++ settings.gradle
@@ -1,1 +1,4 @@
 rootProject.name = 'JasperExporter'
+
+include 'fonts'
+include 'service'
